Reworked factions checkers
